Handle Shellies publishing on multiple topics
The Flood devices publish their status as single values across multiple
topics, as opposed to as a JSON object on one topic. To accommodate
that, had to modify the way that topics are mapped to devices so there
could be multiple topics mapped to a single device. Adjusted the other
devices to match the new pattern.

Also centralized the place where we get the device address from the
device object; I noticed there was one other place where we weren't
generating the address the same way as we were on creation.

Updated the Flood itself to read from multiple topics and set its
drivers accordingly.